Started work on #6
Massive chaos still in the code base, not done combing through the
SqlClient code to split it up. Hit some roadblocks as there are a few
public types in the SqlClient namespace which can't be changed to
another namespace so the chosen namespace will likely change again
before release (or the breaking change marked in the code is accepted as
such).

Work started to make it easier for #1 to be implemented as that's
currently not possible without lots of code duplication and this
refactor has to be done anyway.